requirements.txt
-i https://pypi.org/simple
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
debtcollector==2.2.0; python_version >= '3.6'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==1.7.0; python_version < '3.8'
iso8601==0.1.12
keystoneauth1==4.2.1; python_version >= '3.6'
msgpack==1.0.0
netaddr==0.8.0
netifaces==0.10.9
os-service-types==1.7.0
oslo.config==8.3.1; python_version >= '3.6'
oslo.i18n==5.0.0; python_version >= '3.6'
oslo.serialization==4.0.0; python_version >= '3.6'
oslo.utils==4.4.0; python_version >= '3.6'
packaging==20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pbr==5.4.5
pymysql==0.10.0
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-keystoneclient==4.1.0
python-swiftclient==3.10.0
pytz==2020.1
pyyaml==5.3.1
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rfc3986==1.4.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stevedore==3.2.0; python_version >= '3.6'
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
wrapt==1.12.1
zipp==3.1.0; python_version >= '3.6'
